+++ include/modbus/ModbusTCPCore.h
@@ -9,9 +9,9 @@
 /*!    ModbusTCP core functions */
 namespace ModbusTCPCore
 {
-	int readNextData( ost::TCPStream* tcp, std::queue<unsigned char>& qrecv, int max = 100 );
-	int getNextData( ost::TCPStream* tcp, std::queue<unsigned char>& qrecv, unsigned char* buf, int len );
-	ModbusRTU::mbErrCode sendData( ost::TCPStream* tcp, unsigned char* buf, int len );
+	size_t readNextData( ost::TCPStream* tcp, std::queue<unsigned char>& qrecv, int max = 100 );
+	size_t getNextData( ost::TCPStream* tcp, std::queue<unsigned char>& qrecv, unsigned char* buf, size_t len );
+	ModbusRTU::mbErrCode sendData( ost::TCPStream* tcp, unsigned char* buf, size_t len );
 }
 // -------------------------------------------------------------------------
 #endif // ModbusTCPCore_H_
